// apps/web/src/components/shared/form-builder/types.ts
export interface FormStructure {
  title?: string;
  description?: string;
  orientation?: "vertical" | "horizontal";
  includeHeader?: boolean;
  fieldsets: Fieldset[];
}

export interface Fieldset {
  title?: string;
  description?: string;
  includeHeader?: boolean;
  rows: FieldsetRow[];
}

export interface FieldsetRow {
  className?: string;
  fields: Field[];
}

export enum FieldVariant {
  TEXT = "text",
  EMAIL = "email",
  TEL = "tel",
  NUMBER = "number",
  URL = "url",
  PASSWORD = "password",
  DATE = "date",
  SELECT = "select",
  MULTI_SELECT = "multi_select",
  CHECKBOX = "checkbox",
  CHECK = "check",
  RADIO = "radio",
  SWITCH = "switch",
  TEXTAREA = "textarea",
  IMAGE = "image",
  IMAGE_GALLERY = "image_gallery",
  FILE = "file",
  EMPTY = "empty",
  CUSTOM = "custom",
}

// eslint-disable-next-line @typescript-eslint/no-explicit-any
export interface Field<T=any> {
  id: string;
  label?: string;
  className?: string;
  wrapperClassName?: string;
  variant: FieldVariant;
  required?: boolean;
  description?: string;
  placeholder?: string;
  hidden?: boolean;
  error?: string;
  props?: T;
}

export interface BaseFieldProps {
  disabled?: boolean;
}

export interface TextFieldProps extends BaseFieldProps {
  value?: string | null;
  onChange?: (e: string) => void;
}
// ...
export interface DateFieldProps extends BaseFieldProps {
  value?: Date | null;
  onDateChange?: (e: Date | null) => void;
  nullable?: boolean;
}

export interface SelectOption {
  label: string;
  value: string | number;
}

export interface SelectFieldProps extends BaseFieldProps {
  value?: string | null;
  onValueChange?: (value: string) => void;
  options?: SelectOption[];
}
// ...
export interface TextareaFieldProps extends BaseFieldProps {
  value?: string;
  onChange?: (e: string) => void;
  cols?: number;
  rows?: number;
  resizable?: boolean;
}